refactor: Product를 item으로 네이밍 변경
